fix: make CLI experience for `rocks new` much better
